requirements.txt
playwright
openai
pandas
openpyxl
python-dotenv
argparse
rich
tqdm